power: Handle Boost:INTERACTIVE.
The interactive mode means the device state, e.g. display on or off.
The Boost:INTERACTIVE is used to react on user input e.g. via touch screen.

This was formerly handled on the interaction power hint, which isn't
applicable here. So move this to the boost section and declare it as
supported.

Change-Id: I723618ed9be653d0add8549f810705ee4915c999